switch to gosigar for procstat provider on linux
gosigar performs better than gopsutil. An intrinsic limitation is to be done
with [processes] collection in 0.2s otherwise the listing is incomplete and/or
partially stalled. gosigar does the job at least faster but collection with it
is in full on linux only. gopsutil should suffice on freebsd and darwin. If not,
increase the procstat_ostent update interval (if default 10s is that bad
performance wise) with `--input-interval-procstat-ostent` flag. Eventually there
will be an option to shut procstat_ostent completely off (as with any other
input plugin).